Fix YouTube Chromecast artwork updates.
Unlike most RemoteControlClients, YouTube while Chromecasting sends a large number of onClientMetadataUpdate calls, most without any artwork attached. This fixes us so that we ignore onClientMetadataUpdate calls that do not contain artwork (instead using whatever previous artwork we have).

This doesn't change the behavior of well adapted RemoteControlClients that send a null artwork Bitmap (signifying that they have no artwork for the current track) or the behavior that clears the artwork when the last RemoteControlClient unregisters.