Null check Music BitmapRegionLoader on GL thread
Rare race condition happened when the queued music BitmapRegionLoader was not null prior to adding the callback to the GL thread but null when the GL thread callback happened (which could occur if the music changed in between). As future callbacks will handle the new state, we can early return from the callback if it is null during the callback.