Ensure GLPicture loading does not freeze on low blur amounts
On some devices such as the Nexus 10 and Samsung Galaxy Note 2014, when you use a low amount of blur (triggering the mMaxPrescaledBlurPixels == 0 case), blurrer.destroy() freezes the app, causing images to no longer load until you force stop the app. As the ImageBlurrer is not actually needed unless mMaxPrescaledBlurPixels is greater than 0, we can move the allocation and destruction of it inside the else block. This ensures that the app works as expected in either case.