Convert Advanced Settings Layout to GridLayout
GridLayouts are the new, more performant alternative to TableLayout (as explored in #3). This diff converts the advanced settings layout to use a GridLayout, allowing text to align to the right properly and grow the width of the left column as the text changes (as would happen in different languages).

This uses some extremely lightweight Space elements with columnSpan="2" to add spacing between rows (rather than applying a marginTop to an element and aligning as was done in the RelativeLayout). To keep the number of views low, however, the checkbox rows that only a single element still use marginTop