Allow external application to retrieve the current Artwork details and Bitmap via a ContentProvider
Instead of external applications needing to re-implement all of Muzei or redownload images that Muzei has already cached, we provide a ContentProvider with a single table and single row which stores the current Artwork details and implements openFile() to give access to the already cached artwork Bitmap. MuzeiContract has been added to the api library as well as helper methods for serializing/deserializing Artwork with ContentProvider classes.