Remove Wearable.API warning messages when API is unavailable
When connectionResult.getErrorCode() == ConnectionResult.API_UNAVAILABLE, this means we are on an unsupported device and we should not attempt to update the data layer. However, since that status is unlikely to change over the course of the app's lifecycle, creating a warning message each and every time this happens is unnecessary. Instead, we can do a simple verbose statement that does not spam logcat unless specifically requested.