Improve Watch Face first run experience
In the case where the user has just updated to Muzei 2.0+ (i.e., went from no Wearable app to a Wearable app) AND has not yet had their wallpaper change, there may be cases where the DataLayer has valid artwork, but it is not yet cached locally or made available via the Artwork API.

In those cases, the Muzei watch face would show a notification stating that the user should activate Muzei when no artwork is found. Instead, we should attempt to retrieve the artwork from the DataLayer first and only if that fails should we show the activate Muzei notification.

Note: this moves a lot of code from MuzeiWearableListenerService to ArtworkCacheIntentService, but the functionality is identical.