BackupAgent for advanced settings.
This BackupAgent will backup the advanced settings stored in the default
SharedPreferences. Backing up the default preferences is a bit hacky
since the name isn't easily accessible. Ideally, this would be moved
over to a named SharedPreferences so future OS updates don't break the
backing naming.

Also, need a proper way to handle the backup api key so it isn't stored
in the source code.