Moves base theme to Theme.AppCompat.NoActionBar
As of AppCompat 22.1.0, themes that do not want to use an action bar must contain both
    <item name="windowActionBar">false</item>
    <item name="windowNoTitle">true</item>

Which is automatically done for us when you extend Theme.AppCompat.NoActionBar. As all of our themes, use a Toolbar and not the provided action bar, we can switch our base theme to Theme.AppCompat.NoActionBar and remove the windowActionBar=false declarations scattered throughout our themes.