Update minSdkVersion to 19, move artwork storage to MuzeiProvider
Updates the Artwork table to store multiple rows and store the artwork directly via calls to openOutputStream(), etc. This necessitates moving the minSdkVersion to 19 as I don't know of any way to do what I need to do using Jelly Bean APIs.

This also changes things so that notifyChange() will only be called when the artwork's binary file has been written successfully, rather than when the new artwork information is initially available. This change will ripple through the entire app and be addressed in subsequent commits.