Don't assume the source ComponentName is valid
When inserting sources on Android Wear from the phone/tablet, the ComponentName is, in most cases, not going to be valid. We really only care if it is valid in order to find if we need to send it network available calls, so we can assume if it isn't valid, we won't need to send it network available calls.