Handle Wear sync race condition between source insert and artwork
When syncing to Wear, the artwork sync task may come in slightly before the Source sync comes in. As there is a foreign key requirement that the source row must exist before the artwork row can be inserted, we'll insert a dummy row for the source if it doesn't exist. Expect that this will be corrected when the source sync happens.