Separate notification and Wearable updates from RealRenderController
Instead of putting updating the notification and Wearable on the critical path for updating the wallpaper, create separate threads to handle each case so that all three can be updated in parallel. This should result in much faster wallpaper changes and bind the notification/Wearable updates to the MuzeiWallpaperService lifecycle (as you'd expect).