Filter out duplicate artwork, delete duplicates
Image URI is a soft unique ID of artwork: artwork which have the same non-null image URI are considered to point to the same image. This allows us to filter out duplicate rows in MuzeiProvider by looking at the image URI and only taking the latest row that uses that image URI (noting that an empty image URI is always a unique underlying image).

This somewhat complicates the deleteDocument() case though - since the user only sees the latest artwork row for a given image URI, a delete on that row really means 'delete all artwork with that image URI'.