Add a DATE_ADDED column to Artwork
Change the default sort order for Artwork to use the new DATE_ADDED column (which for sorting purposes, should be identical to _ID), but allows apps to extract actually useful information on when the artwork was updated.

Also allows us to sort the artwork in MuzeiDocumentsProvider by DATE_ADDED and show when the artwork was added.