Make Quick Settings Tile long press open Muzei
When the 'Next Artwork' quick settings tile is unavailable or the user needs to find out more information about Muzei, the correct action is to launch Muzei. There they'll see that the 'Next Artwork' command is not available (no icon shown) or be offered to activate it with more context.