Improve posting and dismiss logic for Activate Muzei notifications
Fixes the Install or Activate notification reappearing multiple times in a row (every time a background is requested). Also fixes the notification not being cleared when Muzei is successfully activated

Also ensures that the Install notification is replaced with the Activate notification if Muzei is installed on the phone but is not actually activated.