Remove invalid tree URIs from the Gallery
Tree URIs can also become invalid, causing SecurityExceptions when accessing them. As updateMetadata is the very first place we try to read the number of photos (and therefore, traverse all of the tree URIs), we can add a check there to find out which URIs we no longer have access to and can therefore delete from the database.