Rotate images before sending to Wear based on Exif Data
Check the ExifInterface.TAG_ORIENTATION for wallpapers and pre-rotate the image before sending to Android Wear. This ensures that the images are always right side up.

Has the unfortunate side effect for requiring a temporary file prior to Android 7.0 due to ExifInterface only supporting files on earlier versions of Android.