Lower the minSdkVersion of the API module to 14
Allow apps to add a dependency on Muzei's API without raising their minSdkVersion beyond API 11, the lowest level the Muzei APIs would technically work on.

Of course, the Muzei app itself has always had a minSdkVersion of at least 17, so the whole API is basically a no-op before that. This just avoids apps using Muzei as a secondary feature having to add an override library to include the dependency.