Remove pending job check from complication scheduling
When calling scheduleComplicationUpdateJob from within onStartJob, the job is still marked as pending. This meant that the pending job check made it so that the job was only scheduled once, rather than repeatedly every time the job is started. As JobScheduler automatically replaces pending jobs with the newly scheduled job, there's no downside to scheduling it every time.