Fix blurring destination Allocation
ImageBlurrer attempts to reuse an Allocation when the Bitmap used to create it was recycled (by MuzeiBlurRenderer). Creating an Allocation wraps the Bitmap, rather than creating a copy so we can create a new Allocation for each new destination Bitmap, preventing reuse abuse.

Moves the source Bitmap input to ImageBlurrer to the constructor to make it clear that there is always one source image (passing in a different image of a different size would have failed anyways). Also fixed an issue where the ScriptIntrinsicColorMatrix was not destroyed.