Add long press on artwork to pin the widget on Android O devices
Android O allows you to pin a widget from within your app. Let users long press the artwork detail screen to bring up the pin the widget screen on supported devices.

This is done asynchronously as to provide a realistic widget preview that includes the actual artwork being shown and whether or not the next artwork button is shown. This makes for a seamless transition when placing the preview on the home screen.

Understandably, putting this behind a long press does not necessarily help with discoverability, but the widget is definitely a very optional piece of Muzei. Future work should be done to find a more discoverable surface for this functionality.