Fix Gallery deletion/force logic for folders
When forcing a folder, we should pick an image from that folder to force, rather than try to force the tree URI itself.

We now use the token of the artwork as the chosen photo URI: this allows us to correctly swap to another image when the folder is removed.