Move Notification Settings to Choose Source overflow menu
The new wallpaper notification setting wasn't really an effect related setting like all the rest on the Advanced screen, but instead is tied to the loading of new artwork (controlled by the Choose Source settings).

Ideally, this should later be moved to a dedicated settings panel that would include more than just it.