Add a Single Artwork source, use it for "Use As..." operations
Avoid cases where the user specifically sets a wallpaper from another app via the "Use As..." operation and then the artwork changes because of Gallery's refresh schedule.

This also reduces the visibility of the GalleryDatabase, ChosenPhoto, and ChosenPhotoDao as they are now purely implementation details of the Gallery source rather than a public API that the main module relies on.