Use the new MuzeiArtProvider API for loading artwork
Route all artwork updates through the new MuzeiArtProvider API, ensuring that only completely loaded artwork makes it into the MuzeiProvider. This removes the case where invalid images are added to the database (sorry Grumpy McPuzzles).

Existing sources are loaded and managed via the SourceArtProvider, which seeks to translate the old API into the new to ensure compatibility.

Completely reworked the Wear app to take advantage of the new API via the new DataLayerArtProvider as it has never talked directly to third party sources.

Also updates the example-source-500px project to give an example of how a native MuzeiArtProvider might look.