Replace ContentObserver on CONTENT_URI with LiveData Observer
Previously the only signal that the artwork's binary data was loaded was the CONTENT_URI change notification. With the move to MuzeiArtProvider, Artwork is only added to the database when it is fully ready. This allows us to convert ContentObservers into LiveData Observers on the Artwork directly.

This also allows us to switch most calls to BitmapRegionLoader to use direct artwork content URIs and not the general CONTENT_URI, avoiding race conditions as the artwork changes out from underneath BitmapRegionLoader.