Allow users to select providers on Wear OS
By allowing users to select which provider they want to use, users are no longer reliant on having Muzei installed on their phone. This opens the ability for other third party apps to publish their app to Wear OS and be selected by Muzei.

Note that commands (outside of the DataLayerArtProvider's Open on phone action and the built in Next Artwork action) are not yet supported.