Adds a way to intercept openFile calls even if the image is cached
When the image is already cached, `openFile(Artwork)` is never called. While that's totally fine for loading the image, in some cases, the MuzeiArtProvider might not know if an image is still valid until right when it is loaded (i.e., the source image could have been deleted).

Add a isArtworkValid method that is called by openFile(Uri, String) to give MuzeiArtProviders the hook needed to invalidate images that are already cached.