Skip disabling DataLayerArtProvider from CapabilityListenerService
The list of nodes can be empty not only when Muzei is uninstalled, but also when the watch loses connectivity to the phone - even in simple cases like the watch going into airplane mode.

We shouldn't disable the From Phone source in those cases. This has the negative side effect that after actually uninstalling Muzei on their phone, they'll still have the From Phone source selected, but that is preferable to the current state of the world and easily fixed by users.