Move SourceManager logic in a separate bound service
Instead of having a single SourceManager that both observes the current provider *and* manages the state of the Sources database, split this functionality into two separate classes.

The new LegacySourceManager is what the MuzeiWallpaperService handles the lifecycle transitions associated with the wallpaper and provider state changes. Only when the wallpaper is active and the legacy source is selected does it bind to the new LegacySourceService.

It is the LegacySourceService that now encapsulates the interactions with the legacy sources including subscribing, unsubscribing, and send the 'Next Artwork' action.