Directly observe WallpaperActiveState to refresh widget
Ensure that any change to the WallpaperActiveState results in the widget being updated. This ensures that the widget shows the 'Next' button when the wallpaper is re-enabled. We no longer need the onDestroy() update because we can use the wallpaper state changing to false to do the the same thing.