Use EXTRA_CHANGED_COMPONENT_NAME_LIST to selectively start artwork loads
Instead of unconditionally starting an artwork load when a matching package changes, only start an artwork load when the selected provider itself changes or when the whole package changes.

This avoids cases where an unrelated setComponentEnabled() change would cause this to fire. This had a significant endless loop when using "Don't Advance" for Auto Advance and was using a Muzei provided source since the next artwork Worker would be the last Worker to run, causing the RescheduleReceiver to be disabled, which would cause an artwork load which would cause the next artwork Worker to run and on and on and on.