Show at most one action in the Wear OS MuzeiActivity
Now that we can use a `RemoteActionCompat` instance, we have access to icons that can be used to show actions from any source installed on the Wear OS device.

Convert the `DataLayerArtProvider` to use `getCommandActions()` and move its "Open on Phone" action to the new structure.

This removes the need for a specific `sendAction` Kotlin extension as we can now rely solely on triggering the `PendingIntent` associated with the command.