Make the ambient time float above the other content
Rather than have the time as part of the scrolling content, it should be always visible along the top when the screen is in ambient mode. This simplifies the relationship between the time and image as now the time has a opaque background and can overlay any content rather than mess with visibility.