Fix leaking MuzeiWallpaperEngine
Since we unconditionally add the `MuzeiWallpaperEngine` as an observer on the `MuzeiWallpaperService` when the engine is created, we need to unconditionally remove the observer when the engine is destroyed.

This fixes leaks when users repeatedly go to the wallpaper preview while Muzei is running in the background where now we properly clean up references to the wallpaper preview engine correctly.

Every little bit helps with #31