Use Call.await() to ensure cooperative cancellation of the call
The okhttp `Call` class supports cancellation. Rather than `execute()` it inline, we use the `await()` helper from `ru.gildor.coroutines:kotlin-coroutines-okhttp` to ensure that cancellation of the coroutine scope managed by WorkManager correctly cancels any inflight Call.

Note that we can't do the same for the call to `string()` as that is a blocking call and okio does not have a non-blocking version as per [this issue](https://github.com/square/okio/issues/501).

Fixes #722