Use a NavigationRailView for MainFragment in landscape
Rather than have a `BottomNavigationView` take the full width of the screen while in landscape (i.e., when the width is greater than the height), swap to a `NavigationRailView` along the left (start) edge of the screen, thus taking up a smaller amount of screen space.

Due to issues in ViewBinding, this required specific casting of `binding.navBar` to the common superclass of `NavigationBarView`.