Move access MuzeiDatabase after device unlock
Since the wallpaper can't actually change until the device is unlocked anyways, we don't need to read from the database at all until the device is unlocked.

This works around https://issuetracker.google.com/issues/432061449 where calling `invalidationTracker.addObserver` (which we do in `MuzeiDatabase.getInstance`) tries to open the database, which it can't do until the device is unlocked.