Collapse Add FAB when items are selected
This commit adds a `LaunchedEffect` to the `GallerySettingsActivity` that observes the number of selected items in the gallery.

When the selection count becomes greater than zero, the "Add" Floating Action Button (FAB) is automatically hidden by setting the `addMode` to `AddNone`. This prevents the user from trying to add new items while in multi-selection mode. When the selection is cleared, the FAB reappears.